Use (s)size_t where appropiate, remove non-unix ifdefs
Drop explicit support for Windows, which is an unrealistic target either way
(no testing for developers, current code would not have worked anyway,
compiling on Windows is a thing for itself).

+++ src/node-v0.10/filter-array.cpp
@@ -23,12 +23,12 @@ namespace lzma {
 FilterArray::FilterArray(Local<Array> arr) : ok_(false) {
 	HandleScope scope;
 	
-	uint32_t len = arr.IsEmpty() ? 0 : arr->Length();
+	size_t len = arr.IsEmpty() ? 0 : arr->Length();
 	
 	Local<String> id_ = String::NewSymbol("id");
 	Local<String> options_ = String::NewSymbol("options");
 	
-	for (uint32_t i = 0; i < len; ++i) {
+	for (size_t i = 0; i < len; ++i) {
 		Local<Object> entry = Local<Object>::Cast(arr->Get(i));
 		if (entry.IsEmpty() || entry->IsUndefined()) {
 			ThrowException(Exception::TypeError(String::New("Filter array needs object entries")));
