+++ src/node-v0.11/filter-array.cpp
@@ -24,12 +24,12 @@ FilterArray::FilterArray(Local<Array> arr) : ok_(false) {
 	Isolate* isolate = Isolate::GetCurrent();
 	HandleScope scope(isolate);
 	
-	uint32_t len = arr.IsEmpty() ? 0 : arr->Length();
+	size_t len = arr.IsEmpty() ? 0 : arr->Length();
 	
 	Local<String> id_ = String::NewFromUtf8(isolate, "id");
 	Local<String> options_ = String::NewFromUtf8(isolate, "options");
 	
-	for (uint32_t i = 0; i < len; ++i) {
+	for (size_t i = 0; i < len; ++i) {
 		Local<Object> entry = Local<Object>::Cast(arr->Get(i));
 		if (entry.IsEmpty() || entry->IsUndefined()) {
 			isolate->ThrowException(Exception::TypeError(String::NewFromUtf8(isolate, "Filter array needs object entries")));
